A lot of updates.
For auto-training.py:
a new technique called auto-training utilises PCA and posterior likelihood to perform clustering of identifiable species to be used as trainign sets for the SVM application of PSYTRANS.

For psytrans:
Script is being modified to input a fixed amount of training data, with the remaining data to be used as testing (within the training set)
Options been put in place to allow for blast result input, auto-training via PCA or the original psytrans procedure to input reference genomes and perform a BLAST classification.
This version also utilises both of the forward and complementary DNA starnd for kmer mapping.

Working but not final version.